move some models to the models folder
