+++ frontend/app/api/Matches.ts
@@ -1,4 +1,4 @@
-import { Match } from '../Match';
+import { Match } from '../models/Match';
 import * as moment from 'moment-timezone';
 import { authHeaders, callApi, fetchArray, maybeFetchObject } from './util';
 import { CreateMatchData } from '../models/CreateMatchData';
