+++ frontend/app/components/host/index.tsx
@@ -5,8 +5,8 @@ import { CreateMatchForm } from './CreateMatchForm';
 import { connect } from 'react-redux';
 import { Dispatch } from 'redux';
 import { nextAvailableSlot } from './nextAvailableSlot';
-import { Regions } from '../../Regions';
-import { renderTeamStyle, TeamStyles } from '../../TeamStyles';
+import { Regions } from '../../models/Regions';
+import { renderTeamStyle, TeamStyles } from '../../models/TeamStyles';
 import { MatchesApi, ApiErrors } from '../../api';
 import { storage } from '../../storage';
 import { change, getFormValues, SubmissionError } from 'redux-form';
