+++ frontend/app/components/upcoming-matches-page/index.tsx
@@ -3,7 +3,7 @@ import { ApplicationState } from '../../state/ApplicationState';
 import { Dispatch } from 'redux';
 import * as React from 'react';
 import { createSelector } from 'reselect';
-import { Match } from '../../Match';
+import { Match } from '../../models/Match';
 import { MatchListing } from '../match-listing';
 import { RouteComponentProps } from 'react-router';
 import { Title } from '../Title';
