+++ frontend/app/models/ApplicationError.ts
