+++ frontend/app/models/BanEntry.ts
