+++ frontend/app/models/TeamStyles.ts
